WeatherApp without UI, only background.

+++ app/src/main/java/com/example/weatherapp/domain/models/Sys.kt
@@ -3,7 +3,7 @@ package com.example.weatherapp.domain.models
 data class Sys(
     val country: String,
     val id: Int,
-    val sunrise: Int,
-    val sunset: Int,
-    val type: Int
+    val sunrise: Long,
+    val sunset: Long,
+    val type: Int,
 )
